requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
requests>=2.28.0
pytest>=7.4.0
hypothesis>=6.82.0
python-dotenv>=1.0.0
schedule>=1.2.0
